Add 2 phase execution to orchest webserver and api
This commits adds the TwoPhaseExecutor (2pe) and the TwoPhaseFunction
class. It introduces the concept of (trying) to do transactional db
updates first, and collateral effects later. This is to have more
control over what's happening, easier debugging, easier testing, etc.
Each TwoPhaseFunction has to implement a transaction and collateral
function, the 2pe, as a context, will take care of running all the
collateral effects when it closes, as well as committing or rollbacking
in case of errors.