Use `logger.info` instead of `echo`
The fact that the global configuration file could not be loaded is more
of an internal concern, than something the user should be aware of.

Consequently, this change fixes the edge case where the very first
time `./orchest start` is invoked (after a clean installation) would
display the error message. This is unwanted behavior since every new
user would be greeted with the error message.