Load views through dispatch
This is necessary to ensure that a change of view will be made with the
up to date global state. This is done by changing the view through the
dispatch mechanism. This necessity was highlighted by the fact that
setting "unsavedChanges" to true and then changing the view could
lead to a user being warned/requested to confirm.