Use cypress-pipe to avoid focus trap issues
Note, this is a somewhat experimental and very specific solution to the
focus trap issue, that makes use of an attempted solution for a similar
problem that Cypress has, which is clicking on objects that are detached
from the DOM, because the application (e.g. react) has rerendered and
those objects have pretty much become garbage. The focus trap issue is
related more to react rather than Cypress, but it seems to have similar
properties.

This is the link to the DOM detachment Cypress github issue
https://github.com/cypress-io/cypress/issues/7306