Delete job pipeline run directory with celery task
Makes it so that the deletion of a job pipeline run directory is done by
the celery worker rather than the orchest webserver. This allows to
avoid a bidirectional relationship between the orchest-webserver and the
orchest-api. The logic related to max_retained_pipeline_runs has been
refactored into a 2PF, DeleteNonRetainedJobPipelineRuns.