Allow abortion of session.launch
In the pre-k8s this wasn't possible, but it opened up the possibility of
a session.launch (thus a interactive session or a job run) getting stuck
due to unforeseen issues. Now that we have migrated to k8s there are
more unknowns so I have added this for safety.