Add DELETE_BASE_IMAGES_CACHE to settings
Note: this is pretty much a placeholder since we don't exactly know in
what other ways the base images cache interacts with updates, etc.
Meaning that the lifecycle of the base images cache is still open for
discussion. It might make more sense to move this button into the
environment images page when that's reworked.