Tune rabbitqm probes
Way more time has been given to rabbitmq to startup, and the health
check in use has been made the most simple as possible. The reason is
that it was producing a good amount of false positives, which led to
rabbitmq being restared, celery and/or the orchest-api failing etc.
The slowness in the starttime is especially evident when cpu is scarce.
More startup time has been given to services relying on rabbitmq.

https://blog.rabbitmq.com/posts/2020/08/deploying-rabbitmq-to-kubernetes-whats-involved/
https://groups.google.com/g/rabbitmq-users/c/6Z_HCO1ENgk