On kernel start inject user defined env vars
As a consequence, the session no longer needs to be restarted to get
updated environment variables. Instead, a kernel restart is enough. This
is a better UX for users since a kernel restart is more lightweight than
a complete session restart.

Caveat: With this change, different kernels (of different steps) can now
have different values for their environment variables. This can be
confusing to users.

Does the caveat outweigh the improved UX?