OrchestSettings.save requires flask app
Previously it was optional. The reason for this change is that save is
currently called in very few places and the cost of change is small,
moreover, and more importantly, the flask app being optional on save
could lead to inconsistent state between the content of the app config
and the settings.