Docs: enable `sphinx-hoverxref` on all references
This is a continuation of the work from #1039 since it was reported that there were
some tooltips that were showing empty content.

This commit will enable hoverxref on all the references of the documentation so
we can debug it deeply and understand why there some cases where the content
comes empty.