On `orchest patch` don't wait on cluster to reach 'status'
Depending on the current cluster's state, `orchest patch` should not
wait for the cluster's state to reach the same state again. For example,
when in RUNNING state then users expect `orchest patch` to succeed once
the cluster is in a RUNNING state again. However, when the cluster is in
an ERROR state then we don't know what the user wants to wait for.

Therefore, we hint the user towards using `orchest status --wait=...`