fix: apply selectors properly to zustand hooks
The return value of the selector shouldn't be wrapped in an array or an
object, as they are new instances, which will always cause re-render.
Selector should return the value as-is for zustand to shallow compare.

Also extract Accordion from EnvironmentAccordion, as we'd also need to
use Accordion in other places.