Tweak DNS resolver `timeout` and `attempts`
Through `dnsConfig` we update the `options` value in `/etc/resolv.conf`.

NOTE: in order to respect the updated options, we need `kube-proxy` to
not update the `iptables` rules to `REJECT` when the `CoreDNS` fails its
`readinessProbe`. Since the DNS settings are not managed by Orchest
directly, we add a note to the doc about this.