Rework image build workflow into pod
This commit does two things:
- it reworks the argo workflow into a pod. The reason is that the
containers that argo injects share the volumes of the container we
specify, and, when using bidirectional mounting, it leads to errors
because said mounting requires a privileged container, and it's
currently not possible to specify such an option for the containers that argo injects. We ended up not needing to use bidirectional mounting at the builder level but I've decided to still go with this since it would take time to revert and we might need bidirectional mountin at the builder level in the future.
- adds support for a buildkit worker by changing what worker is used
based on the container runtime and adjusting the logging sidecar

In hindsight the commit should probably have been split in two.