fix: load wrong job when switching projects
useJob didn't ensure project_uuid is correct and directly use
job_uuid from the URL and result in loading a job that is not in the
current project.

And in fact, it's not necessary to fetchOne job in the summary, as
in useJobsApi state.jobs already include all jobs in the project.
Create useActiveJob hook to replace useJob.