fix: DataTable type errors and some others.
Data for prop "rows" of a DataTable should be wrapped by
DataTableRow, which forces adding "id" to each row.
For most of the use cases, the data from BE has "uuid" instead of "id".
It seems a bit redundant. But this is not in the scope of this PR.
And could be cleaned up later.